+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 05/11/2020
+
+- Remove filter dependency for `expire-cache` endpoint
+
+## 04/11/2020
+
+- Added tests for expire-cache endpoint
+
 # v1.0.0
 
 ## 09/04/2020